+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- [Oban.Migrations] Explicitly set `id` as a `bigserial` to avoid mistakenly
+  generating a `uuid` primary key.
+
 ## [0.2.0] - 2019-05-15
 
 ### Added